requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.8
itemadapter==0.9.0
json5==0.9.25
lxml==5.3.0
python-dotenv==1.0.1
requests==2.32.3
scrapeops_python_logger==0.4.7
scrapeops_python_requests==0.4.7
soupsieve==2.6
tld==0.13
tqdm==4.66.5
urllib3==2.2.3
